+++ jdaviz/configs/imviz/plugins/line_profile_xy/__init__.py
@@ -1,0 +1,1 @@
+from .line_profile_xy import *  # noqa